reorganize and packetize kg code
