+++ kg/dvc.lock
@@ -1,7 +1,7 @@
 schema: '2.0'
 stages:
   create-dataset:
-    cmd: python pipe/create_dataset.py
+    cmd: python src/pipe/create_dataset.py
     deps:
     - path: inputs/arxiv-metadata-oai-snapshot.json
       md5: 00b1b2d81516db5b87c4af81bc2d9f06
@@ -12,7 +12,7 @@ stages:
     - path: inputs/papers-with-abstracts.json
       md5: 553834169476a4ef4452f61f00bfbe81
       size: 436693842
-    - path: pipe/create_dataset.py
+    - path: src/pipe/create_dataset.py
       md5: e307aeddc1f4fb95c642744bbfc7931a
       size: 7953
     outs:
@@ -20,12 +20,12 @@ stages:
       md5: fff7b0ccdb3a126b7507e86961934476
       size: 228998543
   compute-vectors:
-    cmd: python pipe/compute_vectors.py
+    cmd: python src/pipe/compute_vectors.py
     deps:
     - path: assets/dataset.pkl.gz
       md5: fff7b0ccdb3a126b7507e86961934476
       size: 228998543
-    - path: pipe/compute_vectors.py
+    - path: src/pipe/compute_vectors.py
       md5: b5df23787fe55a26370aed1f6c8ff239
       size: 5541
     outs:
@@ -33,7 +33,7 @@ stages:
       md5: ddfccedb23ba5d02a9bcc006ffdeff9e
       size: 3128088704
   check-inputs:
-    cmd: python pipe/check_inputs.py
+    cmd: python src/pipe/check_inputs.py
     deps:
     - path: inputs/arxiv-metadata-oai-snapshot.json
       md5: 00b1b2d81516db5b87c4af81bc2d9f06
@@ -44,18 +44,18 @@ stages:
     - path: inputs/papers-with-abstracts.json
       md5: 553834169476a4ef4452f61f00bfbe81
       size: 436693842
-    - path: pipe/check_inputs.py
+    - path: src/pipe/check_inputs.py
       md5: a636c77345c15d355f79c44671eca49d
       size: 3094
   compute-topics:
-    cmd: python pipe/compute_topics.py
+    cmd: python src/pipe/compute_topics.py
     deps:
     - path: assets/dataset.pkl.gz
       md5: fff7b0ccdb3a126b7507e86961934476
       size: 228998543
     - path: inputs/topics_model
       md5: 6fdd56fa496047913eee00bef84393c1.dir
-    - path: pipe/compute_topics.py
+    - path: src/pipe/compute_topics.py
       md5: a0ef2fab224ef4c2123ff3f9a03202a9
     outs:
     - path: assets/topics.pkl.gz
