+++ kg/src/pipe/check_inputs.py
