+++ kg/src/pipe/compute_vectors.py
@@ -4,10 +4,9 @@ import numpy as np
 import pandas as pd
 import torch
 import transformers
+from pipe.data_checks.vectors import check_vector_size
 from transformers import AutoModel, AutoTokenizer
 from transformers.tokenization_utils_base import BatchEncoding
-
-from data_checks.vectors import check_vector_size
 
 # Paths
 ASSET_PATH = "assets/"
