+++ kg/src/pipe/data_checks/vectors.py
