+++ kg/src/server/tools/assets.py
