+++ kg/tests/pipe/conftest.py
@@ -1,5 +1,5 @@
 import sys
 from pathlib import Path
 
-pipe = Path(Path.cwd(), 'pipe')
+pipe = Path(Path.cwd(), 'src/pipe')
 sys.path.append(str(pipe))
