+++ kg/tests/server/conftest.py
@@ -1,5 +1,5 @@
 import sys
 from pathlib import Path
 
-server = Path(Path.cwd(), 'server')
+server = Path(Path.cwd(), 'src/server')
 sys.path.append(str(server))
